feat: [node-structure-cleanup-003] - Shorten visible graph duration labels
