+++ ui/src/i18n/graph-duration.ts
@@ -1,0 +1,130 @@
+type GraphDurationUnit =
+  | "millennium"
+  | "century"
+  | "year"
+  | "week"
+  | "day"
+  | "hour"
+  | "minute"
+  | "second";
+
+const GRAPH_DURATION_LABELS: Record<string, Record<GraphDurationUnit, string>> = {
+  en: {
+    millennium: "ky",
+    century: "c",
+    year: "y",
+    week: "w",
+    day: "d",
+    hour: "h",
+    minute: "m",
+    second: "s",
+  },
+  "zh-CN": {
+    millennium: "千年",
+    century: "世纪",
+    year: "年",
+    week: "周",
+    day: "天",
+    hour: "时",
+    minute: "分",
+    second: "秒",
+  },
+  ja: {
+    millennium: "千年",
+    century: "世紀",
+    year: "年",
+    week: "週",
+    day: "日",
+    hour: "時",
+    minute: "分",
+    second: "秒",
+  },
+  ko: {
+    millennium: "천년",
+    century: "세기",
+    year: "년",
+    week: "주",
+    day: "일",
+    hour: "시",
+    minute: "분",
+    second: "초",
+  },
+};
+
+const GRAPH_DURATION_UNITS: ReadonlyArray<{ millis: number; unit: GraphDurationUnit }> = [
+  { unit: "millennium", millis: 1000 * 365 * 24 * 60 * 60 * 1000 },
+  { unit: "century", millis: 100 * 365 * 24 * 60 * 60 * 1000 },
+  { unit: "year", millis: 365 * 24 * 60 * 60 * 1000 },
+  { unit: "week", millis: 7 * 24 * 60 * 60 * 1000 },
+  { unit: "day", millis: 24 * 60 * 60 * 1000 },
+  { unit: "hour", millis: 60 * 60 * 1000 },
+  { unit: "minute", millis: 60 * 1000 },
+  { unit: "second", millis: 1000 },
+];
+
+const BASE_GRAPH_DURATION_UNITS = GRAPH_DURATION_UNITS.filter(
+  ({ unit }) => unit === "hour" || unit === "minute" || unit === "second",
+);
+
+export function formatGraphDurationToken(
+  durationMillis: number,
+  locale: string,
+): string {
+  const baseGraphUnit = getBaseGraphDurationUnit(durationMillis);
+  const fittingGraphUnit = getFittingGraphDurationUnit(
+    durationMillis,
+    baseGraphUnit,
+    locale,
+  );
+  const graphValue = Math.floor(durationMillis / fittingGraphUnit.millis);
+
+  return `${new Intl.NumberFormat(locale).format(graphValue)}${getGraphDurationLabel(fittingGraphUnit.unit, locale)}`;
+}
+
+function getBaseGraphDurationUnit(durationMillis: number): {
+  millis: number;
+  unit: GraphDurationUnit;
+} {
+  for (const graphUnit of BASE_GRAPH_DURATION_UNITS) {
+    if (durationMillis >= graphUnit.millis) {
+      return graphUnit;
+    }
+  }
+
+  return BASE_GRAPH_DURATION_UNITS.at(-1) ?? { unit: "second", millis: 1000 };
+}
+
+function getFittingGraphDurationUnit(
+  durationMillis: number,
+  baseUnit: { millis: number; unit: GraphDurationUnit },
+  locale: string,
+): { millis: number; unit: GraphDurationUnit } {
+  const baseUnitIndex = GRAPH_DURATION_UNITS.findIndex(
+    ({ unit }) => unit === baseUnit.unit,
+  );
+
+  for (let index = baseUnitIndex; index >= 0; index -= 1) {
+    const graphUnit = GRAPH_DURATION_UNITS[index];
+    if (!graphUnit) {
+      continue;
+    }
+    const graphValue = Math.floor(durationMillis / graphUnit.millis);
+    if (graphValue <= getMaxGraphDurationValue(graphUnit.unit, locale)) {
+      return graphUnit;
+    }
+  }
+
+  return GRAPH_DURATION_UNITS[0];
+}
+
+function getMaxGraphDurationValue(
+  unit: GraphDurationUnit,
+  locale: string,
+): number {
+  const maxDigits = Math.max(1, 3 - getGraphDurationLabel(unit, locale).length);
+  return 10 ** maxDigits - 1;
+}
+
+function getGraphDurationLabel(unit: GraphDurationUnit, locale: string): string {
+  return GRAPH_DURATION_LABELS[locale]?.[unit] ?? GRAPH_DURATION_LABELS.en[unit];
+}
